Redefined module optimizer #160
