+++ src/packages/piral-cli-parcel/src/parcel/bundler.ts
@@ -3,26 +3,13 @@ import extendBundlerWithAtAlias = require('parcel-plugin-at-alias');
 import extendBundlerWithCodegen = require('parcel-plugin-codegen');
 import extendBundlerWithImportMaps = require('parcel-plugin-import-maps');
 import { PiletSchemaVersion } from 'piral-cli';
+import { log } from 'piral-cli/utils';
 import { extendBundlerWithExternals, combineExternals } from 'parcel-plugin-externals/utils';
 import { SourceMapConsumer, SourceMapGenerator } from 'source-map';
 import { existsSync, statSync, readFile, writeFile } from 'fs';
 import { resolve, dirname, basename } from 'path';
-import { patchModule } from './bundler-patches';
 import { extendConfig } from './settings';
 import { BundlerSetup } from '../types';
-import {
-  log,
-  computeHash,
-  checkExists,
-  checkIsDirectory,
-  readJson,
-  readText,
-  createFileIfNotExists,
-  writeText,
-  writeJson,
-  getFileNames,
-  ForceOverwrite,
-} from 'piral-cli/utils';
 
 let original: any;
 
@@ -90,82 +77,6 @@ export function gatherJsBundles(
   }
 
   return gatheredBundles;
-}
-
-// See https://github.com/smapiot/piral/issues/121#issuecomment-572055594
-const defaultIgnoredPackages = ['core-js'];
-
-/**
- * The motivation for this method came from:
- * https://github.com/parcel-bundler/parcel/issues/1655#issuecomment-568175592
- * General idea:
- * Treat all modules as non-optimized for the current output target.
- * This makes sense in general as only the application should determine the target.
- */
-async function patch(staticPath: string, ignoredPackages: Array<string>) {
-  log('generalDebug_0003', `Patching files in "${staticPath}" ...`);
-  const folderNames = await getFileNames(staticPath);
-  return Promise.all(
-    folderNames.map(async folderName => {
-      if (!ignoredPackages.includes(folderName)) {
-        const rootName = resolve(staticPath, folderName);
-        const isDirectory = await checkIsDirectory(rootName);
-
-        if (isDirectory) {
-          if (folderName.startsWith('@')) {
-            // if we are scoped, just go down
-            await patch(rootName, ignoredPackages);
-          } else {
-            try {
-              const packageFileData = await readJson(rootName, 'package.json');
-
-              if (packageFileData.name && packageFileData._piralOptimized === undefined) {
-                packageFileData._piralOptimized = packageFileData.browserslist || true;
-                delete packageFileData.browserslist;
-
-                await writeJson(rootName, 'package.json', packageFileData, true);
-                await writeText(rootName, '.browserslistrc', 'node 10.11');
-                await patchModule(folderName, rootName);
-              }
-
-              await patchFolder(rootName, ignoredPackages);
-            } catch (e) {
-              log('generalDebug_0003', `Encountered a patching error: ${e}`);
-            }
-          }
-        }
-      }
-    }),
-  );
-}
-
-async function patchFolder(rootDir: string, ignoredPackages: Array<string>) {
-  const file = '.patched';
-  const modulesDir = resolve(rootDir, 'node_modules');
-  const exists = await checkExists(modulesDir);
-
-  if (exists) {
-    const lockContent = (await readText(rootDir, 'package-lock.json')) || (await readText(rootDir, 'yarn.lock'));
-    const currHash = computeHash(lockContent);
-    const prevHash = await readText(modulesDir, file);
-    log('generalDebug_0003', `Evaluated patch module hashes: "${currHash}" and "${prevHash}".`);
-
-    if (prevHash !== currHash) {
-      await patch(modulesDir, ignoredPackages);
-      await createFileIfNotExists(modulesDir, file, currHash, ForceOverwrite.yes);
-    }
-  }
-}
-
-export async function patchModules(rootDir: string, ignoredPackages = defaultIgnoredPackages) {
-  log('generalDebug_0003', `Patching modules starting in "${rootDir}" ...`);
-  const otherRoot = resolve(require.resolve('parcel-bundler'), '..', '..', '..');
-  await patchFolder(rootDir, ignoredPackages);
-
-  if (otherRoot !== rootDir) {
-    log('generalDebug_0003', `Also patching modules in "${otherRoot}" ...`);
-    await patchFolder(otherRoot, ignoredPackages);
-  }
 }
 
 const bundleUrlRef = '__bundleUrl__';
