+++ src/packages/piral-cli-parcel/src/parcel/run-debug-piral.ts
@@ -1,6 +1,6 @@
 import { LogLevels } from 'piral-cli';
-import { setStandardEnvs, progress } from 'piral-cli/utils';
-import { setupBundler, patchModules } from './bundler';
+import { setStandardEnvs } from 'piral-cli/utils';
+import { setupBundler } from './bundler';
 
 async function run(
   root: string,
@@ -45,11 +45,6 @@ process.on('message', async msg => {
   switch (msg.type) {
     case 'bundle':
       if (bundler) {
-        if (msg.optimizeModules) {
-          progress('Preparing modules ...');
-          await patchModules(root, msg.ignored);
-        }
-
         await bundler.bundle();
 
         bundler.on('buildStart', () => {
