+++ src/packages/piral-cli/src/common/index.ts
@@ -18,6 +18,7 @@ export * from './merge';
 export * from './npm';
 export * from './pack';
 export * from './package';
+export * from './patcher';
 export * from './patches';
 export * from './port';
 export * from './rules';
